backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
python-multipart
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
tabulate==0.9.0